+++ aqme/csearch/__init__.py
@@ -1,0 +1,1 @@
+from aqme.csearch.base import csearch
