+++ aqme/qprep.py
@@ -1,3 +1,45 @@
+"""
+Parameters
+----------
+   files : mol object, str or list of str, default=None
+      This module prepares input QM file(s). Formats accepted: mol object(s), 
+      Gaussian or ORCA LOG/OUT output files, JSON, XYZ, SDF, PDB. Also, 
+      lists can be used (i.e. [FILE1.log, FILE2.log] or \*.FORMAT such as \*.json).
+   atom_types : list of str, default=[]
+      (If files is None) List containing the atoms of the system
+   cartesians : list of str, default=[]
+      (If files is None) Cartesian coordinates used for further processing
+   w_dir_main : str, default=os.getcwd()
+      Working directory
+   destination : str, default=None,
+      Directory to create the input file(s)
+   varfile : str, default=None
+      Option to parse the variables using a yaml file (specify the filename)
+   program : str, default=None
+      Program required to create the new input files. Current options: 'gaussian', 'orca'
+   qm_input : str, default=''
+      Keywords line for new input files (i.e. 'B3LYP/6-31G opt freq')
+   qm_end : str, default=''
+      Final line(s) in the new input files
+   charge : int, default=None
+      Charge of the calculations used in the following input files. If charge isn't defined, it defaults to 0
+   mult : int, default=None
+      Multiplicity of the calculations used in the following input files. If mult isn't defined, it defaults to 1
+   suffix : str, default=''
+      Suffix for the new input files (i.e. FILENAME_SUFFIX.com for FILENAME.log)
+   chk : bool, default=False
+      Include the chk input line in new input files for Gaussian calculations
+   mem : str, default='4GB'
+      Memory for the QM calculations (i) Gaussian: total memory; (ii) ORCA: memory per processor
+   nprocs : int, default=2
+      Number of processors used in the QM calculations
+   gen_atoms : list of str, default=[]
+      Atoms included in the gen(ECP) basis set (i.e. ['I','Pd'])
+   bs_gen : str, default=''
+      Basis set used for gen(ECP) atoms (i.e. 'def2svp')
+   bs_nogen : str, default=''
+      Basis set used for non gen(ECP) atoms in gen(ECP) calculations (i.e. '6-31G*')
+"""
 ######################################################.
 #        This file stores the QPREP class            #
 ######################################################.
@@ -17,7 +59,7 @@ from aqme.utils import (
     read_xyz_charge_mult,
     mol_from_sdf_or_mol_or_mol2,
 )
-from aqme.csearch_crest_utils import xyzall_2_xyz
+from aqme.csearch.crest import xyzall_2_xyz
 from pathlib import Path
