+++ docs/API/API_Reference.rst
@@ -1,0 +1,27 @@
+.. automodule:: aqme
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Main Modules
+============
+
+.. toctree::
+   :maxdepth: 1
+
+   aqme.csearch
+   aqme.cmin
+   aqme.qprep
+   aqme.qcorr
+   aqme.qdescp
+
+Other Modules
+=============
+
+.. toctree::
+   :maxdepth: 1
+
+   aqme.qcorr_utils
+   aqme.utils
+   aqme.vismol
+
