+++ docs/API/aqme.cmin.rst
@@ -1,0 +1,5 @@
+cmin
+====
+
+.. automodule:: aqme.cmin
+   :members:
