+++ docs/API/aqme.csearch.rst
@@ -1,0 +1,12 @@
+csearch
+=======
+
+.. autoclass:: aqme.csearch.csearch
+   :members:
+
+.. automodule:: aqme.csearch.base
+
+
+
+
+
