+++ docs/API/aqme.filter.rst
@@ -1,0 +1,5 @@
+filter
+======
+
+.. automodule:: aqme.filter
+   :members:
