+++ docs/API/aqme.qcorr.rst
@@ -1,0 +1,7 @@
+qcorr
+=====
+
+.. autoclass:: aqme.qcorr.qcorr
+   :members:
+
+.. automodule:: aqme.qcorr
