+++ docs/API/aqme.qdescp.rst
@@ -1,0 +1,8 @@
+qdescp
+======
+
+.. autoclass:: aqme.qdescp.qdescp
+   :members:
+
+.. automodule:: aqme.qdescp
+
