+++ docs/API/aqme.qprep.rst
@@ -1,0 +1,8 @@
+qprep
+=====
+
+.. autoclass:: aqme.qprep.qprep
+   :members:
+
+.. automodule:: aqme.qprep
+
