+++ docs/API/aqme.utils.rst
@@ -1,0 +1,5 @@
+utils
+=====
+
+.. automodule:: aqme.utils
+   :members:
