+++ docs/API/aqme.vismol.rst
@@ -1,0 +1,8 @@
+vismol
+======
+
+.. autoclass:: aqme.vismol.vismol
+   :members:
+
+.. automodule:: aqme.vismol
+
