+++ docs/Generation of Gaussian Input Files/Options for input parameters.rst
@@ -1,5 +1,5 @@
 .. _Options for input parameters:
 
-===============
+============================
 Options for input parameters
-===============
+============================
