+++ docs/Misc/license.rst
@@ -1,4 +1,8 @@
-.. _license:
-
+=======
 License
 =======
+
+.. include:: ../../README.rst
+   :start-after: license-start
+   :end-before: license-end
+
