+++ docs/Misc/versions.rst
@@ -4,5 +4,21 @@
 Versions
 ========
 
-Version 1.0.0
-   - First release of AQME!
+Version 1.3.1 [`url <https://github.com/jvalegre/aqme/releases/tag/1.3.1>`__]
+   -  Workflows were updated
+   -  Small fixes in CREST when using constraints
+   -  Readme was updated
+   -  GoodVibes added in installation requirements
+
+Version 1.3.0 [`url <https://github.com/jvalegre/aqme/releases/tag/1.3.0>`__]
+   -  Publication version
+
+Version 1.2.0 [`url <https://github.com/jvalegre/aqme/releases/tag/1.2.0>`__]
+   -  This version improves how AQME reads PATHs from arguments to make the program more robust
+
+Version 1.1.0 [`url <https://github.com/jvalegre/aqme/releases/tag/1.1.0>`__]
+   -  Fixes pip install issue coming from older versions
+
+
+Version 1.0.0 [`url <https://github.com/jvalegre/aqme/releases/tag/1.0.0>`__]
+   -  First official version of AQME ready to generate publication-quality results
