+++ docs/Quickstart/basic.rst
@@ -1,43 +1,10 @@
-.. _basic:
+=====
+Usage
+=====
 
-===============
-Basic Structure
-===============
-
-This documents provides steps for running the ``AQME`` program.
-
-.. contents::
-
-Command line
-------------
-
-The ``AQME`` program can be run from the command line by providing the necessary arguments. In the following example,
-an input file is provided and the compute option in invoked.
-
-.. code:: bash
-
-  $ python -m aqme --input [file] --compute
-
-
-YAML file
----------
-
-The ``AQME`` program can also be run by using a yaml file which contains the necessary arguments. In the following example,
-a yaml file named ``params.yaml`` is created where an input file name is provided and the compute option in invoked.
-
-.. code-block:: yaml
-
-  # INPUT FILE
-  input : 'file.smi' # input files
-
-  # COMPUTE OPTION TRUE
-  compute : True
-
-To run ``AQME`` program with the yaml file, the following command is used
-
-.. code:: bash
-
-  $ python -m aqme --varfile params.yaml
+.. include:: ../../README.rst
+   :start-after: quickstart-start
+   :end-before: quickstart-end
 
 
 Additional parameters which are needed to be changed can be added in the yaml file. A complete YAML file is
