+++ docs/Quickstart/requirements.rst
@@ -1,35 +1,4 @@
-.. _requirements:
+.. include:: ../../README.rst
+   :start-after: requirements-start
+   :end-before: requirements-end
 
-============
-Requirements
-============
-
-The following packages would be required for successfully running ``AQME``!
-
-python
-------
-
-- python 3.7
-- ase (pip install ase==3.19.1)
-- shutil (pip install pytest-shutil)
-- glob (pip install glob2==0.4.1)
-- pandas (pip install pandas==1.0.3)
-- yaml (pip install PyYAML==5.3.1)
-- progress (pip install progress==1.5)
-
-Programs for Optimization
--------------------------
-
-- xtb
-- ANI1ccx
-- torch (pip install torch)
-- torchvision (pip install torchvision)
-- torchani (pip install torchani==2.0)
-
-Cheminformatics
----------------
-
-- RDkit (conda install -c rdkit rdkit=2020.03.2.0)
-- Openbabel (conda install -c openbabel openbabel==2.4.1)
-- cclib (conda install -c conda-forge cclib)
-- goodvibes (pip install goodvibes==3.0.1)
