+++ docs/conf.py
@@ -10,9 +10,11 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+# Ensure that modules can be imported without installing aqme
+sys.path.insert(0, os.path.abspath('..')) 
+
 
 
 # -- Project information -----------------------------------------------------
@@ -22,7 +24,7 @@ copyright = '2020, Shree Sowndarya S. V., Juan V. Alegre Requena, Robert S. Pato
 author = 'Shree Sowndarya S. V., Juan V. Alegre Requena, Robert S. Paton'
 
 # The full version, including alpha/beta/rc tags
-release = 'v1.1'
+release = 'v1.3'
 
 
 # -- General configuration ---------------------------------------------------
@@ -30,8 +32,12 @@ release = 'v1.1'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = [
+extensions = ['sphinx.ext.autodoc',
 ]
+
+# Avoid paths in class names i.e. 
+# class aqme.aqme.csearch.csearch -> class csearch
+add_module_names = False
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -51,4 +57,4 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+#html_static_path = ['_static']
